Move getIntroductoryOfferIntervalDisplay to wpcom-checkout
