+++ packages/wpcom-checkout/src/index.ts
@@ -22,3 +22,4 @@ export * from './payment-methods/existing-credit-card';
 export { isWpComProductRenewal } from './is-wpcom-product-renewal';
 export { isValueTruthy } from './is-value-truthy';
 export * from './checkout-labels';
+export * from './get-introductory-offer-interval-display';
